+++ src/task2/Entity/Diamond.java
@@ -3,10 +3,10 @@ package task2.Entity;
 import task2.Calculate.Calculation;
 
 public class Diamond extends Figure implements Calculation {
-    private int diagonal1;
-    private int diagonal2;
+    private double diagonal1;
+    private double diagonal2;
 
-    public Diamond(Colors background,Colors border,int diagonal1,int diagonal2) {
+    public Diamond(Colors background,Colors border,double diagonal1,double diagonal2) {
         super(FigureType.DIAMOND,background,border);
         this.diagonal1=diagonal1;
         this.diagonal2=diagonal2;
