+++ src/task2/Entity/Hexagon.java
@@ -3,8 +3,8 @@ package task2.Entity;
 import task2.Calculate.Calculation;
 
 public class Hexagon extends Figure implements Calculation {
-    private int radius;
-    public Hexagon(Colors background,Colors border,int radius) {
+    private double radius;
+    public Hexagon(Colors background,Colors border,double radius) {
         super(FigureType.HEXAGON,background,border);
         this.radius=radius;
     }
