+++ src/task2/Entity/Octagon.java
@@ -3,9 +3,9 @@ package task2.Entity;
 import task2.Calculate.Calculation;
 
 public class Octagon extends Figure implements Calculation {
-    private int radius;
+    private double radius;
 
-    public Octagon(Colors background,Colors border,int radius) {
+    public Octagon(Colors background,Colors border,double radius) {
         super(FigureType.OCTAGON,background,border);
         this.radius=radius;
     }
