+++ src/task2/Entity/Oval.java
@@ -3,12 +3,12 @@ package task2.Entity;
 import task2.Calculate.Calculation;
 
 public class Oval extends Figure implements Calculation {
-    private int radius1;
-    private int radius2;
-    private int radius3;
+    private double radius1;
+    private double radius2;
+    private double radius3;
 
 
-    public Oval(Colors background,Colors border,int radius1,int radius2,int radius3) {
+    public Oval(Colors background,Colors border,double radius1,double radius2,double radius3) {
         super(FigureType.OVAL,background,border);
         this.radius1=radius1;
         this.radius2=radius2;
