+++ src/task2/Entity/Rectangle.java
@@ -4,9 +4,9 @@ import task2.Calculate.Calculation;
 
 public class Rectangle extends Figure implements Calculation {
 
-    private int side1;
-    private int side2;
-    public Rectangle(Colors background,Colors border,int side1,int side2) {
+    private double side1;
+    private double side2;
+    public Rectangle(Colors background,Colors border,double side1,double side2) {
         super(FigureType.RECTANGLE,background,border);
         this.side1=side1;
         this.side2=side2;
