+++ src/task2/Entity/Square.java
@@ -3,9 +3,9 @@ package task2.Entity;
 import task2.Calculate.Calculation;
 
 public class Square extends Figure implements Calculation {
-    private int side;
+    private double side;
 
-    public Square( Colors background, Colors border,int side) {
+    public Square( Colors background, Colors border,double side) {
         super(FigureType.SQUARE, background, border);
         this.side=side;
     }
