+++ src/task2/Entity/Triangle.java
@@ -3,10 +3,10 @@ package task2.Entity;
 import task2.Calculate.Calculation;
 
 public class Triangle extends Figure implements Calculation {
-    private int side1;
-    private int side2;
-    private int side3;
-    public Triangle(Colors background,Colors border,int side1,int side2,int side3) {
+    private double side1;
+    private double side2;
+    private double side3;
+    public Triangle(Colors background,Colors border,double side1,double side2,double side3) {
         super(FigureType.TRIANGLE,background,border);
         this.side1=side1;
         this.side2=side2;
