+++ src/task2/Factory/OctagonFactory.java
@@ -6,7 +6,7 @@ import task2.Entity.Figure;
 import task2.Entity.Octagon;
 
 public interface OctagonFactory {
-    default Figure create(Colors background, Colors border, int radius) {
+    default Figure create(Colors background, Colors border, double radius) {
         return new Octagon(background,border,radius);
 
     }
