+++ src/task2/Factory/OvalFactory.java
@@ -6,7 +6,7 @@ import task2.Entity.Figure;
 import task2.Entity.Oval;
 
 public interface OvalFactory{
-    default Figure create(Colors background, Colors border, int radius1,int radius2,int radius3) {
+    default Figure create(Colors background, Colors border, double radius1,double radius2,double radius3) {
         return new Oval(background,border,radius1,radius2,radius3);
 
     }
