+++ src/task2/Factory/RectangleFactory.java
@@ -4,7 +4,7 @@ import task2.Entity.*;
 
 public interface RectangleFactory {
 
-        default Figure create(Colors background, Colors border, int side1,int side2) {
+        default Figure create(Colors background, Colors border, double side1,double side2) {
             return new Rectangle(background,border,side1,side2);
 
         }
