+++ src/task2/Factory/SquareFactory.java
@@ -6,7 +6,7 @@ import task2.Entity.Square;
 import task2.Entity.Triangle;
 
 public interface SquareFactory {
-    default Figure create(Colors background, Colors border, int side) {
+    default Figure create(Colors background, Colors border, double side) {
         return new Square(background,border,side);
 
     }
