+++ src/task2/Factory/TriangleFactory.java
@@ -5,7 +5,7 @@ import task2.Entity.Figure;
 import task2.Entity.Triangle;
 
 public interface TriangleFactory {
-    default Figure create(Colors background, Colors border, int side1, int side2, int side3) {
+    default Figure create(Colors background, Colors border, double side1, double side2, double side3) {
         return new Triangle(background,border,side1,side2,side3);
 
     }
